Add change grid functionality
We added a button to allow the user to reset the
number of squares in the grid, and gave it an upper
limit of 100, so you wont crash your comp doing it.
If you go past the warning message and put a too large
number into it again, it'll let you. Probbably should be
bound by some "while(isTooBig=true) function but whatever.